requirements.txt
# Phase 1 Baseline Vulnerability Assessment Dependencies

# Core ML/NLP
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # For memory-efficient loading

# Data handling
PyYAML>=6.0
pandas>=2.0.0
numpy>=1.24.0

# Metrics and analysis
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.66.0

# Optional: HuggingFace datasets for additional test prompts
datasets>=2.14.0

# Utilities
python-dateutil>=2.8.0
